+++ doc/releaseNotes.md
@@ -443,3 +443,5 @@ This file describes the main feature changes for each readout.exe released versi
 ## next version
 - readout-monitor: added log file for received measurements.
 - added support for custom commands to be executed on states transitions.
+- Updated configuration parameters:
+  - readout.flushConsumerTimeout: when set, readout waits up to this amount of time that all data pages locked by consumers are released before stopping.